requirements.txt
django
python-magic-bin==0.4.14